s44_backend/requirements.txt
flask
flask-jwt-extended
flask-cors
mysql-connector-python
python-dotenv
gunicorn
pymongo[srv]
Flask-PyMongo